requirements.txt
asyncpraw
mplfinance
nltk
opencv-python
pip-chill
praw
pytesseract
python-dotenv
selenium
tabulate
yfinance
